appointmentservice/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
pika==1.3.1
pytz==2022.6
sqlparse==0.4.3
django-rest-knox
Pillow
tzdata==2022.1
mysql-connector-python-rf